Move catalog backend to impls
